+++ resources/views/admin/newnote.blade.php
@@ -22,8 +22,8 @@ New Note « Admin CP
 @stop
 
 @section('scripts')
-<link rel="styelsheet" href="https://api.mapbox.com/mapbox.js/v2.2.3/mapbox.css">
-<script src="https://api.mapbox.com/mapbox.js/v2.2.3/mapbox.js"></script>
+<link rel="stylesheet" href="https://api.mapbox.com/mapbox.js/v2.2.3/mapbox.css" integrity="sha384-ZDBUvY/seENyR1fE6u4p1oMFfsKVjIlkiB6TrCdXjeZVPlYanREcmZopTV8WFZ0q" crossorigin="anonymous">
+<script src="https://api.mapbox.com/mapbox.js/v2.2.3/mapbox.js" integrity="sha384-GgNQAMOzWcL0fePJqogHh8dCjsGKZpkBNgm3einGr0aUb9kcXvr9JeU/PDf5knja" crossorigin="anonymous"></script>
 
 <script src="/assets/js/newnote.js"></script>
 <script src="/assets/bower/store2.min.js"></script>